common macroquad code in their example
